Do not call euca when credentials are not set.

+++ deploy_stack.py
@@ -72,7 +72,8 @@ def prepare_environment(env, already_bootstrapped, machines):
 
 def destroy_environment(client, instance_tag):
     client.destroy_environment()
-    if client.env.config['type'] == 'manual':
+    if (client.env.config['type'] == 'manual'
+            and 'AWS_ACCESS_KEY' in os.environ):
         destroy_job_instances(instance_tag)
